runs/urls.py
from django.urls import path, include
from rest_framework.routers import DefaultRouter
from .views import RunViewSet

router = DefaultRouter()
router.register(r"runs", RunViewSet, basename="run")

urlpatterns = [
    path("", include(router.urls)),
]
